refactor: package 명칭변경 (comm -> util)
