+++ src/main/java/com/nexon/maple/userInfo/service/UserInfoWriteService.java
@@ -1,20 +1,13 @@
 package com.nexon.maple.userInfo.service;
 
-import com.nexon.maple.comm.maplestoryHomepage.CustomMapleComment;
-import com.nexon.maple.otp.entity.Otp;
 import com.nexon.maple.otp.repository.OtpDao;
-import com.nexon.maple.terms.entity.TermsAgreeInfo;
-import com.nexon.maple.terms.entity.TermsInfo;
 import com.nexon.maple.terms.repository.TermsInfoDao;
-import com.nexon.maple.userInfo.dto.RegisterUserInfoCommand;
 import com.nexon.maple.userInfo.entity.UserInfo;
 import com.nexon.maple.userInfo.repository.UserInfoDao;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.Assert;
-
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
