+++ src/main/java/com/nexon/maple/util/maplestoryHomepage/CustomDocument.java
@@ -1,4 +1,4 @@
-package com.nexon.maple.comm.maplestoryHomepage;
+package com.nexon.maple.util.maplestoryHomepage;
 
 import org.jsoup.Connection;
 import org.jsoup.Jsoup;
