+++ src/main/java/com/nexon/maple/util/maplestoryHomepage/otpComment.java
@@ -1,4 +1,4 @@
-package com.nexon.maple.comm.maplestoryHomepage;
+package com.nexon.maple.util.maplestoryHomepage;
 
 public interface otpComment {
     public boolean equalsComment(String otpNumber);
