+++ src/test/java/com/nexon/maple/util/maplestoryHomepage/CustomDocumentTest.java
@@ -1,4 +1,4 @@
-package com.nexon.maple.comm.maplestoryHomepage;
+package com.nexon.maple.util.maplestoryHomepage;
 
 import org.junit.jupiter.api.Test;
